Renamed app-footer to footer
