--- code/src/layouts/app-footer/messages/index.ts
@@ -1,4 +1,0 @@
-import en from "@/layouts/app-footer/messages/en.json";
-import uk from "@/layouts/app-footer/messages/uk.json";
-
-export default { en, uk };
