+++ code/src/layouts/footer/messages/index.ts
@@ -1,0 +1,4 @@
+import en from "@/layouts/footer/messages/en.json";
+import uk from "@/layouts/footer/messages/uk.json";
+
+export default { en, uk };
