+++ code/src/layouts/root-layout/RootLayout.tsx
@@ -1,13 +1,13 @@
 import { Outlet } from "react-router-dom";
 import Header from "@/layouts/header/Header";
-import AppFooter from "@/layouts/app-footer/AppFooter";
+import Footer from "@/layouts/footer/Footer";
 
 const RootLayout = () => {
   return (
     <>
       <Header />
       <Outlet />
-      <AppFooter />
+      <Footer />
     </>
   );
 };
