+++ code/src/messages.ts
@@ -1,6 +1,6 @@
 import bannerMessages from "@/layouts/intro-banner/messages";
 import commonMessages from "@/constants/common-messages";
-import footerMessages from "@/layouts/app-footer/messages";
+import footerMessages from "@/layouts/footer/messages";
 import bestSellersMessages from "@/layouts/best-sellers/messages";
 import subintroMessages from "@/layouts/subintro/messages";
 import categorySectionMessages from "@/layouts/category-section/messages";
